d/control: depend on polkitd instead of policykit-1
`policykit-1` has recently been split in 2 packages and replaced by a
transitional package. The one we actually need is `polkitd`, so let's
have this one as the "main" dependency.

However, as this package is often backported to bullseye as well, let's
use an alternative as `polkitd` only exists in current testing/sid.

Closes: #1022072.